+++ ReleaseNotes.md
@@ -7,6 +7,13 @@ Eclipse Update Site:
 *   Releases: <https://dl.bintray.com/pmd/pmd-eclipse-plugin/updates/>
 *   Snapshots: <https://dl.bintray.com/pmd/pmd-eclipse-plugin/snapshots/updates/>
 
+## ????: 4.2.1.v????
+
+This is a bugfix release.
+
+### Fixed Issues
+
+
 ## 31-March-2019: 4.2.0.v20190331-1136
 
 This is a minor release.