+++ ReleaseNotes.md
@@ -7,7 +7,7 @@ Eclipse Update Site:
 *   Releases: <https://dl.bintray.com/pmd/pmd-eclipse-plugin/updates/>
 *   Snapshots: <https://dl.bintray.com/pmd/pmd-eclipse-plugin/snapshots/updates/>
 
-## ????: 4.2.1.v????
+## 26-April-2019: 4.2.1.v20190426-1741
 
 This is a bugfix release.